Break up the generic datalad.js resolvers into logical files.
