+++ packages/openneuro-server/graphql/resolvers/snapshots.js
@@ -1,0 +1,13 @@
+import { getSnapshot, getSnapshots } from '../../datalad/snapshots.js'
+import { dataset } from './dataset.js'
+
+export const snapshots = obj => {
+  return getSnapshots(obj.id)
+}
+
+export const snapshot = (obj, { datasetId, tag }, context) => {
+  return getSnapshot(datasetId, tag).then(snapshot => ({
+    ...snapshot,
+    dataset: () => dataset(snapshot, { id: datasetId }, context),
+  }))
+}
